requirements-merged.txt
# ============================================
# Mini-Medallion: Core Dependencies
# ============================================

# --- Data & Numerics ---
numpy>=1.26.0
pandas>=2.1.0
polars>=0.20.0
pyarrow>=14.0.0
scipy>=1.12.0

# --- Machine Learning ---
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.2.0
hmmlearn>=0.3.0

# --- Deep Learning (CPU fallback) ---
torch>=2.2.0
torchvision>=0.17.0

# --- Signal Processing ---
PyWavelets>=1.5.0

# --- Evolutionary Algorithms ---
deap>=1.4.1

# --- Data Sources ---
yfinance>=0.2.36
fredapi>=0.5.1
requests>=2.31.0

# --- Databases & Storage ---
questdb>=1.1.0
redis>=5.0.0
minio>=7.2.0

# --- MLOps ---
mlflow>=2.10.0

# --- Visualization ---
matplotlib>=3.8.0
plotly>=5.18.0
seaborn>=0.13.0

# --- Configuration ---
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# --- Logging & Monitoring ---
loguru>=0.7.0
prometheus-client>=0.20.0

# --- Utilities ---
tqdm>=4.66.0
click>=8.1.0
rich>=13.7.0
schedule>=1.2.0
# ============================================
# Mini-Medallion: Base Requirements (CPU/GPU Agnostic)
# ============================================
# Common dependencies shared across all environments.
# For GPU: pip install -r requirements-gpu.txt
# For CPU: pip install -r requirements-cpu.txt
# For complete setup: pip install -r requirements-base.txt

# --- Configuration & Utilities ---
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
click>=8.1.0

# --- Logging & Monitoring ---
loguru>=0.7.0
prometheus-client>=0.20.0

# --- REST API ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# --- Data Processing (core APIs, no GPU acceleration) ---
numpy>=1.26.0
pandas>=2.1.0
polars>=0.20.0
pyarrow>=14.0.0
scipy>=1.12.0

# --- Machine Learning (CPU-only models) ---
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.2.0
hmmlearn>=0.3.0

# --- Signal Processing ---
PyWavelets>=1.5.0

# --- Evolutionary Algorithms ---
deap>=1.4.1

# --- Data Sources ---
yfinance>=0.2.36
fredapi>=0.5.1
requests>=2.31.0

# --- Scheduling ---
schedule>=1.2.0

# --- Databases & Storage ---
questdb>=1.1.0
redis>=5.0.0
minio>=7.2.0

# --- MLOps ---
mlflow>=2.10.0

# --- Visualization ---
matplotlib>=3.8.0
plotly>=5.18.0
seaborn>=0.13.0

# --- Utilities ---
tqdm>=4.66.0
click>=8.1.0
rich>=13.7.0
schedule>=1.2.0

# --- Development & Testing ---
pytest>=7.4.0
pytest-cov>=4.1.0
